reorganize mock files
