+++ monitor.cfg
@@ -3,6 +3,7 @@ host = 0.0.0.0
 port = 1501
 plugins = openstack_generic
 debug = True
+retries= 0
 
 [openstack_generic]
 key_pair = /home/asperathos/.ssh/asperathos_key
