+++ monitor/tests/mocks/mock_monasca.py
